masked type imports the string not the other way around :)
